kinetics/requirements.txt
numpy==1.19.5
protobuf==3.19.6
tensorflow==2.5.0
dm-sonnet==2.0.2
tensorflow-probability==0.24.0
tensorboard==2.5.0
moviepy==1.0.3
Pillow==10.4.0
tqdm==4.66.5
